Implement comprehensive CVSS v4.0 calculator with official RedHat implementation
- Add cvssV4FullImplementation.ts with official CVSS v4.0 MacroVector algorithm
- Include comprehensive test suite with 52 test cases covering all EQ combinations
- Update cvssV4Calculator.ts to use unified scoring approach
- Remove cvssV4Official.ts in favor of full implementation
- Add BSD-2-Clause license attribution to FIRST.ORG and Red Hat
- Update all existing tests to work with new implementation
- Ensure compatibility with existing CVSSV4Vector interface
- Validate against known vectors and official calculator results
